Add VirtualSubsystem and use it for Vision (#8)
The Subsystem class in Command based has two purposes that can get
muddled in user code. The first is the state mutexing for Commands and
the second is acting as a physical/conceptual subsystem, something with
set code that runs and serves a purpose (moving a mechanism, detecting
objects etc). VirtualSubsystem is for systems which don't need mutexing
but do need code to run periodically such as vision.

Signed-off-by: Jade Turner <spacey-sooty@proton.me>